Move files in the phone bill GWT original project to the "phonebillgwt" package so that we can have unique package names in the original projects that will not confuse the aggregate javadoc build.
