+++ gwt-parent/gwt-originals-parent/phonebill-gwt/src/it/java/edu/pdx/cs410J/phonebillgwt/PingServiceSyncProxyIT.java
@@ -1,8 +1,8 @@
-package edu.pdx.cs410J.original;
+package edu.pdx.cs410J.phonebillgwt;
 
 import com.gdevelop.gwt.syncrpc.SyncProxy;
-import edu.pdx.cs410J.original.client.PhoneBill;
-import edu.pdx.cs410J.original.client.PingService;
+import edu.pdx.cs410J.phonebillgwt.client.PhoneBill;
+import edu.pdx.cs410J.phonebillgwt.client.PingService;
 import edu.pdx.cs410J.web.HttpRequestHelper;
 import org.junit.Test;
