+++ gwt-parent/gwt-originals-parent/phonebill-gwt/src/it/java/edu/pdx/cs410J/phonebillgwt/client/PhoneBillGwtTestSuite.java
@@ -1,4 +1,4 @@
-package edu.pdx.cs410J.original.client;
+package edu.pdx.cs410J.phonebillgwt.client;
 
 import com.google.gwt.junit.tools.GWTTestSuite;
 import junit.framework.Test;
