+++ gwt-parent/gwt-originals-parent/phonebill-gwt/src/main/java/edu/pdx/cs410J/phonebillgwt/client/PingService.java
@@ -1,8 +1,7 @@
-package edu.pdx.cs410J.original.client;
+package edu.pdx.cs410J.phonebillgwt.client;
 
 import com.google.gwt.user.client.rpc.RemoteService;
 import com.google.gwt.user.client.rpc.RemoteServiceRelativePath;
-import edu.pdx.cs410J.AbstractPhoneBill;
 
 /**
  * A GWT remote service that returns a dummy Phone Bill
