+++ gwt-parent/gwt-originals-parent/phonebill-gwt/src/main/java/edu/pdx/cs410J/phonebillgwt/client/PingServiceAsync.java
@@ -1,4 +1,4 @@
-package edu.pdx.cs410J.original.client;
+package edu.pdx.cs410J.phonebillgwt.client;
 
 import com.google.gwt.user.client.rpc.AsyncCallback;
