+++ gwt-parent/gwt-originals-parent/phonebill-gwt/src/main/resources/edu/pdx/cs410J/phonebillgwt/PhoneBillGwt.gwt.xml
@@ -2,7 +2,7 @@
   <inherits name="com.google.gwt.user.User"/>
   <inherits name='com.google.gwt.user.theme.standard.Standard'/>
   <inherits name='edu.pdx.cs410J.Projects'/>
-  <entry-point class="edu.pdx.cs410J.original.client.PhoneBillGwt"/>
+  <entry-point class="edu.pdx.cs410J.phonebillgwt.client.PhoneBillGwt"/>
 
-  <servlet path="/ping" class="edu.pdx.cs410J.original.server.PingServiceImpl"/>
+  <servlet path="/ping" class="edu.pdx.cs410J.phonebillgwt.server.PingServiceImpl"/>
 </module>
